Updates to Planet Summary
I like the color changes you made, and the see through nature of it,
although you can't click through it.

Moved your new settings for the Slider to the myJslider

Changed from GridLayout to Gridbaglayout. so that we can better
customize the layout.

Created a .lock function in the myJslider, it's demonstrated with
research.

I also put the buttons on, but we'll probably want to change to
JCheckBox.